refactor(cmds): extract shared sidecar core; dedupe ratls-mesh CDS upgrade
Extract internal/cmds/sidecar with the CDS-release plumbing shared by
get-secret and get-volume (config, flags, validation, mTLS client,
challenge/sandbox-token request, retry loop); get-secret's hand-rolled
atomic write becomes fileutil.WriteAtomic.

Share ratls-mesh's host and in-guest CDS cert-upgrade goroutines as
runCDSUpgrade, and replace its local parseHexMeasurements with
ratls.ParseHexMeasurements, both already covered by pkg tests.

Also adds docs/improvement-plan.md, the audit this work executes.
