+++ internal/cmds/getsecret/cmd.go
@@ -1,9 +1,9 @@
 package getsecret
 
 import (
-	"time"
-
 	"github.com/spf13/cobra"
+
+	"github.com/confidential-dot-ai/c8s/internal/cmds/sidecar"
 )
 
 // NewCmd returns the cobra subcommand. Registered as a child of `c8s`.
@@ -41,17 +41,9 @@ does not survive a CDS restart; nothing durable may be keyed on it.`,
 		},
 	}
 	f := cmd.Flags()
-	f.StringVar(&cfg.CDSURL, "cds-url", "", "https base URL of CDS")
-	f.StringVar(&cfg.AttestationApiURL, "attestation-api-url", "", "local attestation-api used to verify CDS's RA-TLS certificate")
-	f.StringSliceVar(&cfg.Measurements, "measurements", nil, "SHA-384 hex launch measurement(s) CDS must present (repeatable; empty pins none, UNSAFE)")
-	f.StringVar(&cfg.CertPath, "cert", "/run/c8s/certs/tls.crt", "the pod's CDS-issued certificate, presented to CDS")
-	f.StringVar(&cfg.KeyPath, "key", "/run/c8s/certs/tls.key", "private key for --cert")
+	sidecar.BindFlags(f, &cfg.Config, "per-request timeout against CDS")
 	f.StringSliceVar(&specs, "secret", nil, "NAME=/store/path to fetch; NAME is the filename written under --out-dir (repeatable)")
 	f.StringVar(&cfg.OutDir, "out-dir", "/run/c8s/secrets", "directory the secret files are written to; must be memory-backed")
 	f.StringVar(&cfg.FileMode, "file-mode", "0640", "octal mode for the written files")
-	f.IntVar(&cfg.Attempts, "attempts", 60, "how many times to try before failing; release is refused until every main container is running, so retries are expected")
-	f.DurationVar(&cfg.RetryInterval, "retry-interval", 5*time.Second, "wait between attempts")
-	f.DurationVar(&cfg.RequestTimeout, "request-timeout", 10*time.Second, "per-request timeout against CDS")
-	f.DurationVar(&cfg.InventoryTimeout, "inventory-timeout", 5*time.Second, "timeout for redeeming a sandbox token from the node's admission inventory")
 	return cmd
 }
